requirements.txt
openai
Django
requests
django-model-utils
mysqlclient
gunicorn
jinja2
pytz
python-dateutil
django-jinja
django-bootstrap-v5
django-environ